IMP: ironed out some problems with bundled DCMTK. Still not building JPEG-support libraries.
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@191 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/Utilities/dcmtk/import_from_dcmtk.sh
@@ -36,8 +36,8 @@ copy_src_tgt()
   mkdir -p ${tgt_c_path}
 
   for module in ${module_list}; do
-    cp ${src_h_path}/${module}.h ${tgt_h_path}
-    cp ${src_c_path}/${module}.cc ${tgt_c_path}
+      [ -f ${src_h_path}/${module}.h ] && cp ${src_h_path}/${module}.h ${tgt_h_path}
+      [ -f ${src_c_path}/${module}.cc] && cp ${src_c_path}/${module}.cc ${tgt_c_path}
   done
 
 }
